Use glanceclient to discover images
Getting information about images via novaclient limits a bit
the metadata that's available at image level. One bit that's
especially interesting is the visibility details, as it allows
making sure that snapshots are not considered when looking for
the latest image of a given flavor.

This change requires a new setting in ai.conf (GlanceClient timeout).