(AI-4822) - ensure ip(6)tables in running early
Fixes AI-4822 - After the first reboot puppet would try
to insert iptable rules which failed since the
ip(6)tables service was not running.

Fixing this as a dependency in puppet is not obious since
iptables-save must also be issued after the rules are added.

For CC7 and RHEL7 before first reboot we disable firewalld
and enable iptables and ip6tables so they are running
before puppet tries to insert iptables rule.